Implement precompiles for bw6 plonk proof verification
The precompile is going to be called from a solidity contract whose
sole role will be to pass the buffer with serialized proof, vk and
witness. All the verification is done in the precompile using the ususal
Gnark API.

The precompile lives at the address of 0x85, which is the next available
address in our collection of precompiles. The gas cost was estimated at
21000, although it is not backed by any serious calculations.

Signed-off-by: Wojciech Zmuda <zmuda.w@gmail.com>