PlonK proof verifier precompile. for BW6-761 (#6)
* Update Gnark to v0.9.2

Since Quorum deserializes objects generated by the in-circuit verifier,
both projects must use the same Gnark version to ensure binary
compatibility. Update Gnark to the same version the in-circuit verifier
uses.

Signed-off-by: Wojciech Zmuda <zmuda.w@gmail.com>

* Implement precompiles for bw6 plonk proof verification

The precompile is going to be called from a solidity contract whose
sole role will be to pass the buffer with serialized proof, vk and
witness. All the verification is done in the precompile using the ususal
Gnark API.

The precompile lives at the address of 0x85, which is the next available
address in our collection of precompiles. The gas cost was estimated at
21000, although it is not backed by any serious calculations.

Signed-off-by: Wojciech Zmuda <zmuda.w@gmail.com>

* Apply suggestions from code review

Co-authored-by: Ara Adkins <iamrecursion@users.noreply.github.com>

---------

Signed-off-by: Wojciech Zmuda <zmuda.w@gmail.com>
Co-authored-by: Ara Adkins <iamrecursion@users.noreply.github.com>